Added existential types and fixed a number of problems with pattern
matching.

+++ compiler/compiler.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015, Jay Conrod. All rights reserved.
+# Copyright 2014-2016, Jay Conrod. All rights reserved.
 #
 # This file is part of Gypsum. Use of this source code is governed by
 # the GPL license that can be found in the LICENSE.txt file.
@@ -9,7 +9,7 @@ from functools import partial
 import ast
 from bytecode import W8, W16, W32, W64, BUILTIN_TYPE_CLASS_ID, BUILTIN_TYPE_CTOR_ID, instInfoByCode, BUILTIN_MATCH_EXCEPTION_CLASS_ID, BUILTIN_MATCH_EXCEPTION_CTOR_ID, BUILTIN_STRING_EQ_OP_ID
 from ir import IrTopDefn, Class, Field, Function, Global, LOCAL, Package, PACKAGE_INIT_NAME, RECEIVER_SUFFIX, Variable
-from ir_types import NoType, UnitType, BooleanType, I8Type, I16Type, I32Type, I64Type, F32Type, F64Type, ClassType, VariableType, NULLABLE_TYPE_FLAG, getExceptionClassType, getClassFromType, getStringType, getRootClassType
+from ir_types import Type, NoType, UnitType, BooleanType, I8Type, I16Type, I32Type, I64Type, F32Type, F64Type, ClassType, VariableType, ExistentialType, NULLABLE_TYPE_FLAG, getExceptionClassType, getClassFromType, getStringType, getRootClassType
 import ir_instructions
 from compile_info import CONTEXT_CONSTRUCTOR_HINT, CLOSURE_CONSTRUCTOR_HINT, PACKAGE_INITIALIZER_HINT, ARRAY_ELEMENT_GET_HINT, ARRAY_ELEMENT_SET_HINT, ARRAY_ELEMENT_LENGTH_HINT, DefnInfo, NORMAL_MODE, STD_MODE, NOSTD_MODE
 from flags import ABSTRACT, STATIC, LET, ARRAY
@@ -615,11 +615,13 @@ class CompileVisitor(ast.NodeVisitor):
 
     def visitTupleExpression(self, expr, mode):
         ty = self.info.getType(expr)
+        typeArgs = map(self.info.getType, expr.expressions)
         tupleClass = ty.clas
         langMode = self.info.languageMode()
         assert langMode is not NOSTD_MODE
 
         # Allocate the tuple.
+        self.buildStaticTypeArguments(typeArgs)
         if langMode is NORMAL_MODE:
             self.allocobjf(tupleClass)
         else:
@@ -1061,15 +1063,28 @@ class CompileVisitor(ast.NodeVisitor):
         assert self.isDetached()
 
     def visitPartialFunctionCase(self, expr, mode, ty, doneBlock, failBlock):
-        if expr.condition is not None:
+        if expr.condition is None:
+            self.visit(expr.pattern, COMPILE_FOR_MATCH, ty, failBlock)
+        else:
+            # If there's a condition, duplicate the value before matching the pattern. The
+            # pattern will consume the value on a successful match, but we need to preserve
+            # it if the condition is false.
             assert self.unreachable or failBlock is not None
             failBlockId = failBlock.id if failBlock is not None else -1
             self.dup()
-        self.visit(expr.pattern, COMPILE_FOR_MATCH, ty, failBlock)
-        if expr.condition is not None:
+            mustMatch = type_analysis.patternMustMatch(expr.pattern, ty, self.info)
+            if mustMatch:
+                self.visit(expr.pattern, COMPILE_FOR_MATCH, ty, None)
+            else:
+                dropBlock = self.newBlock()
+                self.visit(expr.pattern, COMPILE_FOR_MATCH, ty, dropBlock)
             self.visit(expr.condition, COMPILE_FOR_VALUE)
             successBlock = self.newBlock()
             self.branchif(successBlock.id, failBlockId)
+            if not mustMatch:
+                self.setCurrentBlock(dropBlock)
+                self.drop()
+                self.branch(failBlockId)
             self.setCurrentBlock(successBlock)
             self.drop()
         self.visit(expr.expression, mode)
@@ -1644,57 +1659,58 @@ class CompileVisitor(ast.NodeVisitor):
             self.dup()
             self.buildStaticTypeArguments(typeArgs)
             self.callFunction(irDefn)
+        returnType = irDefn.returnType.substitute(irDefn.typeParameters, typeArgs)
+        optionClass = self.info.getStdClass("Option", loc)
+        optionType = returnType.substituteForBaseClass(optionClass)
+        valueType = optionType.getTypeArguments()[0]
 
         # Check if it returned Some.
-        someClass = self.info.getStdClass("Some", loc)
-        someType = ClassType.forReceiver(someClass)
-        self.buildType(someType, loc)
+        self.dup()
+        self.buildStaticTypeArgument(valueType)
+        self.buildCallNamedMethod(optionType, "is-defined", COMPILE_FOR_VALUE)
         matcherSuccessBlock = self.newBlock()
-        dropSomeBlock = self.newBlock()
-        self.castcbr(matcherSuccessBlock.id, dropSomeBlock.id)
+        dropBlock = self.newBlock()
+        self.branchif(matcherSuccessBlock.id, dropBlock.id)
 
         # Get the value out of the Some. Cast it to a tuple if we need to.
         self.setCurrentBlock(matcherSuccessBlock)
-        self.buildStaticTypeArgument(getRootClassType())
-        self.buildCallNamedMethod(someType, "get", COMPILE_FOR_VALUE)
-        n = len(subPatterns)
-        if n > 1:
-            tupleClass = self.info.getTupleClass(n, loc)
-            tupleType = ClassType.forReceiver(tupleClass)
-            self.buildType(tupleType, loc)
-            self.castc()
+        self.buildStaticTypeArgument(valueType)
+        self.buildCallNamedMethod(optionType, "get", COMPILE_FOR_VALUE)
 
-        # If there is just one pattern inside, pass the value to that pattern. Otherwise,
-        # load each value out of the tuple and pass those to the patterns. Note that we pass
-        # Object as the expression type for each tuple. We haven't properly type checked the
-        # value we pulled out of the Some[_]. We need to type check these values at runtime,
-        # but the type checks should never fail.
+        # If there is just one sub-pattern, pass the value to that pattern. Otherwise, the
+        # value must be a tuple. Load each value out of the tuple and pass those to the
+        # sub-patterns.
         successState = None
+        n = len(subPatterns)
         if n == 1:
-            self.visit(subPatterns[0], COMPILE_FOR_MATCH, getRootClassType(), dropSomeBlock)
+            self.visit(subPatterns[0], COMPILE_FOR_MATCH, valueType, dropBlock)
             successState = self.saveCurrentBlock()
         else:
+            valueTypeArgs = valueType.getTypeArguments()
             dropFieldBlock = self.newBlock()
+            mustMatch = True
             for i in xrange(n - 1):
                 self.dup()
                 self.ldf(i)
-                self.visit(subPatterns[i], COMPILE_FOR_MATCH,
-                           getRootClassType(), dropFieldBlock)
+                self.visit(subPatterns[i], COMPILE_FOR_MATCH, valueTypeArgs[i], dropFieldBlock)
+                mustMatch &= type_analysis.patternMustMatch(subPatterns[i], valueTypeArgs[i],
+                                                           self.info)
             self.ldf(n - 1)
-            self.visit(subPatterns[-1], COMPILE_FOR_MATCH, getRootClassType(), dropSomeBlock)
+            self.visit(subPatterns[-1], COMPILE_FOR_MATCH, valueTypeArgs[-1], dropBlock)
             successState = self.saveCurrentBlock()
 
             # If one of the sub-patterns failed to match, we need to drop the field.
-            self.setCurrentBlock(dropFieldBlock)
-            self.drop()
-            self.branch(dropSomeBlock.id)
+            if not mustMatch:
+                self.setCurrentBlock(dropFieldBlock)
+                self.drop()
+                self.branch(dropBlock.id)
 
         # If the matcher did not return Some, drop the return value.
-        self.setCurrentBlock(dropSomeBlock)
+        self.setCurrentBlock(dropBlock)
         self.drop()
         self.branch(failBlock.id)
 
-        # Go back to the success block and drop the last field.
+        # Go back to the success block and drop value passed to the matcher.
         self.restoreCurrentBlock(successState)
         self.drop()
 
@@ -1718,18 +1734,9 @@ class CompileVisitor(ast.NodeVisitor):
         """Builds a type on the static type argument stack and simultaneously builds an
         equivalent Type object on the value stack. This is useful for checked casts and other
         type tests."""
-        if isinstance(ty, VariableType):
-            if ty.typeParameter.isForeign():
-                self.tyvdf(ty.typeParameter)
-            else:
-                self.tyvd(ty.typeParameter)
-        else:
-            assert isinstance(ty, ClassType)
+        if isinstance(ty, ClassType):
             assert len(ty.clas.typeParameters) == len(ty.typeArguments)
             for param, arg in zip(ty.clas.typeParameters, ty.typeArguments):
-                if STATIC in param.flags and arg != VariableType(param):
-                    raise SemanticException(
-                        loc, "cannot dynamically check a type with static type parameters")
                 self.buildType(arg, loc)
             if ty.clas.isForeign():
                 self.tycdf(ty.clas)
@@ -1737,6 +1744,17 @@ class CompileVisitor(ast.NodeVisitor):
                 self.tycd(ty.clas)
             if ty.isNullable():
                 self.tyflagd(1)
+        elif isinstance(ty, VariableType):
+            if ty.typeParameter.isForeign():
+                self.tyvdf(ty.typeParameter)
+            else:
+                self.tyvd(ty.typeParameter)
+        else:
+            assert isinstance(ty, ExistentialType)
+            for v in ty.variables:
+                self.buildType(VariableType(v), loc)
+            self.buildType(ty.ty, loc)
+            self.tyxd(len(ty.variables))
 
     def buildImplicitStaticTypeArguments(self, typeParams):
         for param in typeParams:
@@ -1761,7 +1779,11 @@ class CompileVisitor(ast.NodeVisitor):
             assert not ty.typeParameter.isForeign()
             self.tyvs(ty.typeParameter)
         else:
-            raise NotImplementedError
+            assert isinstance(ty, ExistentialType)
+            for v in ty.variables:
+                self.buildStaticTypeArgument(VariableType(v))
+            self.buildStaticTypeArgument(ty.ty)
+            self.tyxs(len(ty.variables))
 
     def buildCast(self, ty):
         self.buildStaticTypeArgument(ty)
